add in proper logics for armored and stealth enemies

+++ Tower Defense/Assets/Scripts/GameManager.cs
@@ -35,7 +35,7 @@ public class GameManager : MonoBehaviour
 
     private void HandleEnemyReachedEnd(EnemyData data)
     {
-        _health = Mathf.Max(0, _health - data.damage);
+        _health = (int)Mathf.Max(0, _health - data.damage);
         OnHealthChanged?.Invoke(_health);
     }
